Integrate stats rings into data validation header
